# packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis AI",
    "description": "\n**Aegis AI** \u2014 Continuous AI Assurance & Governance Platform.\n\nContinuously test AI systems and agents for fairness, hallucination, safety, privacy, security and policy\ncompliance \u2014 with evidence-backed findings and automatic re-testing.\n\nAuthentication: session cookie, session bearer token, or an API key (`Authorization: Bearer aeg_live_\u2026`).\nAll data is scoped to the authenticated workspace.\n",
    "contact": {
      "name": "Aegis AI"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/policies": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List Policies",
        "operationId": "list_policies_api_v1_policies_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PolicyOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Create Policy",
        "operationId": "create_policy_api_v1_policies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{policy_id}": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Get Policy",
        "operationId": "get_policy_api_v1_policies__policy_id__get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{policy_id}/versions": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Policy Versions",
        "operationId": "policy_versions_api_v1_policies__policy_id__versions_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyVersionOut"
                  },
                  "title": "Response Policy Versions Api V1 Policies  Policy Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{policy_id}/upload": {
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_policies__policy_id__upload_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_policies__policy_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{policy_id}/compile": {
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Compile Policy",
        "operationId": "compile_policy_api_v1_policies__policy_id__compile_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/policies/{policy_id}/diff": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Diff",
        "operationId": "diff_api_v1_policies__policy_id__diff_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          },
          {
            "name": "from_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "From Version"
            }
          },
          {
            "name": "to_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "To Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDiffOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{policy_id}/controls": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Policy Controls",
        "operationId": "policy_controls_api_v1_policies__policy_id__controls_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ControlOut"
                  },
                  "title": "Response Policy Controls Api V1 Policies  Policy Id  Controls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Add Manual Control",
        "operationId": "add_manual_control_api_v1_policies__policy_id__controls_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualControlCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/controls": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List Controls",
        "operationId": "list_controls_api_v1_controls_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ControlOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/frameworks/{framework_id}": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "Get Framework",
        "operationId": "get_framework_api_v1_frameworks__framework_id__get",
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Framework Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_policies__policy_id__upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_policies__policy_id__upload_post"
      },
...
      "CompileResultOut": {
        "properties": {
          "policy_version_id": {
            "type": "string",
            "title": "Policy Version Id"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementOut"
            },
            "type": "array",
            "title": "Requirements"
          },
          "controls": {
            "items": {
              "$ref": "#/components/schemas/ControlOut"
            },
            "type": "array",
            "title": "Controls"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          }
        },
        "type": "object",
        "required": [
          "policy_version_id",
          "requirements",
          "controls",
          "report"
        ],
        "title": "CompileResultOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualControlCreate": {
        "properties": {
          "control_id": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9]{1,7}-\\d{2,4}$",
            "title": "Control Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "test_type": {
            "$ref": "#/components/schemas/TestType"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "default": "medium"
          },
          "threshold": {
            "additionalProperties": true,
            "type": "object",
            "title": "Threshold"
          },
          "condition": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "required_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Evidence"
          }
        },
        "type": "object",
        "required": [
          "control_id",
          "name",
          "test_type"
        ],
        "title": "ManualControlCreate"
      },
...
      "Page_PolicyOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PolicyOut"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[PolicyOut]"
      },
...
      "PolicyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "key": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^[A-Za-z][A-Za-z0-9]*$",
            "title": "Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text",
            "description": "Policy text to compile into controls"
          }
        },
        "type": "object",
        "required": [
          "name",
          "key"
        ],
        "title": "PolicyCreate"
      },
      "PolicyDiffOut": {
        "properties": {
          "from_version": {
            "type": "string",
            "title": "From Version"
          },
          "to_version": {
            "type": "string",
            "title": "To Version"
          },
          "added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed"
          },
          "changed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changed"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "requires_reaudit": {
            "type": "boolean",
            "title": "Requires Reaudit"
          },
          "affected_controls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Controls"
          }
        },
        "type": "object",
        "required": [
          "from_version",
          "to_version",
          "added",
          "removed",
          "changed",
          "unchanged",
          "requires_reaudit",
          "affected_controls"
        ],
        "title": "PolicyDiffOut"
      },
      "PolicyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key",
          "description",
          "category",
          "status",
          "owner_name",
          "current_version_id",
          "is_demo",
          "created_at",
          "updated_at"
        ],
        "title": "PolicyOut"
      },
...
      "PolicyVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "compiled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compiled At"
          },
          "compiler_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compiler Version"
          },
          "compile_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Compile Report"
          },
          "change_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "status",
          "source_type",
          "compiled_at",
          "compiler_version",
          "compile_report",
          "change_note",
          "created_at"
        ],
        "title": "PolicyVersionOut"
      },
...
    }
  },
...
}